Changed SourceTreeItem from interface to class
