+++ src/main/java/org/roda/rodain/source/ui/items/SourceTreeLoading.java
@@ -1,15 +1,10 @@
 package org.roda.rodain.source.ui.items;
 
-import javafx.scene.control.TreeItem;
-import org.roda.rodain.rules.Rule;
-
-import java.util.Observable;
-
 /**
  * @author Andre Pereira apereira@keep.pt
  * @since 17-09-2015.
  */
-public class SourceTreeLoading extends TreeItem<String> implements SourceTreeItem{
+public class SourceTreeLoading extends SourceTreeItem{
     public SourceTreeLoading(){
         super("Loading...");
     }
@@ -22,33 +17,5 @@ public class SourceTreeLoading extends TreeItem<String> implements SourceTreeIte
     @Override
     public SourceTreeItemState getState(){
         return SourceTreeItemState.NORMAL;
-    }
-
-    @Override
-    public void addIgnore(){
-    }
-
-    @Override
-    public void addMapping(Rule r){
-    }
-
-    @Override
-    public void removeIgnore() {
-
-    }
-
-    @Override
-    public void removeMapping(Rule r) {
-
-    }
-
-    @Override
-    public void forceUpdate() {
-
-    }
-
-    @Override
-    public void update(Observable o, Object arg) {
-
     }
 }
